# services/marketing-ops/openapi/marketing-ops.v1.yaml
openapi: 3.1.0
info:
  title: Nexus Marketing Ops API
  version: 1.0.0
  description: REST contract for the operational campaign workspace.
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
...
  /campaign-items/{itemId}/content-assets:
    get:
      operationId: listProductionItemContentAssets
      summary: List stable content assets for an item
      parameters:
        - { $ref: '#/components/parameters/ItemId' }
      responses:
        '200':
          description: Content assets
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentAssetListEnvelope' }
        default: { $ref: '#/components/responses/Error' }
    post:
      operationId: createProductionItemContentAsset
      summary: Create a stable content asset
      parameters:
        - { $ref: '#/components/parameters/ItemId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentAssetCreate' }
      responses:
        '201':
          description: Content asset created
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentAssetMutationEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /content-assets/{assetId}/versions:
    get:
      operationId: listContentVersions
      summary: List immutable content versions
      parameters:
        - { $ref: '#/components/parameters/AssetId' }
      responses:
        '200':
          description: Immutable content history
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentVersionListEnvelope' }
        default: { $ref: '#/components/responses/Error' }
    post:
      operationId: createContentVersion
      summary: Append a content version
      parameters:
        - { $ref: '#/components/parameters/AssetId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentVersionCreate' }
      responses:
        '201':
          description: Immutable content version created
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentVersionMutationEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /campaign-items/{itemId}/artifacts:
    get:
      operationId: listProductionItemArtifacts
      summary: List active artifact links
      parameters:
        - { $ref: '#/components/parameters/ItemId' }
      responses:
        '200':
          description: Active item artifacts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ItemArtifactListEnvelope' }
        default: { $ref: '#/components/responses/Error' }
    post:
      operationId: attachProductionItemArtifact
      summary: Upload or link an item artifact
      parameters:
        - { $ref: '#/components/parameters/ItemId' }
        - { $ref: '#/components/parameters/FilenameOptional' }
        - { $ref: '#/components/parameters/AssetIdOptional' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ItemArtifactLink' }
          application/pdf:
            schema: { type: string, format: binary, maxLength: 26214400 }
          application/vnd.openxmlformats-officedocument.wordprocessingml.document:
            schema: { type: string, format: binary, maxLength: 26214400 }
          application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
            schema: { type: string, format: binary, maxLength: 26214400 }
          application/vnd.openxmlformats-officedocument.presentationml.presentation:
            schema: { type: string, format: binary, maxLength: 26214400 }
          text/plain:
            schema: { type: string, format: binary, maxLength: 26214400 }
          text/csv:
            schema: { type: string, format: binary, maxLength: 26214400 }
          image/png:
            schema: { type: string, format: binary, maxLength: 26214400 }
          image/jpeg:
            schema: { type: string, format: binary, maxLength: 26214400 }
          image/webp:
            schema: { type: string, format: binary, maxLength: 26214400 }
      responses:
        '201':
          description: Item artifact attached
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ItemArtifactMutationEnvelope' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      operationId: unlinkProductionItemArtifact
      summary: Logically unlink an item artifact
      parameters:
        - { $ref: '#/components/parameters/ItemId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ItemArtifactRemoval' }
      responses:
        '200':
          description: Artifact link removed
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ItemArtifactRemovalEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /campaign-items/{itemId}/artifacts/{artifactLinkId}/access-link:
    post:
      operationId: createProductionItemArtifactAccessLink
      summary: Create a short-lived artifact access link
      parameters:
        - { $ref: '#/components/parameters/ItemId' }
        - { $ref: '#/components/parameters/ArtifactLinkId' }
      responses:
        '200':
          description: Short-lived authorized access link
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MaterialAccessEnvelope' }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  headers:
    ETag:
      description: Quoted current resource or aggregate version.
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
  parameters:
...
    ItemId: { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
    AssetId: { name: assetId, in: path, required: true, schema: { type: string, format: uuid } }
    ArtifactLinkId: { name: artifactLinkId, in: path, required: true, schema: { type: string, format: uuid } }
...
    FilenameOptional: { name: X-Nexus-Filename, in: header, required: false, schema: { type: string, minLength: 1, maxLength: 255 } }
    AssetIdOptional: { name: X-Nexus-Asset-Id, in: header, required: false, schema: { type: string, format: uuid } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 128 } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string, pattern: '^(W/)?("[1-9][0-9]*"|[1-9][0-9]*)$' } }
  responses:
    Error:
      description: Stable public error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    PublicErrorCode:
      type: string
      enum:
        - validation_error
        - unauthorized
        - auth_unavailable
        - tenant_required
        - forbidden
        - not_found
        - origin_forbidden
        - feature_disabled
        - idempotency_key_required
        - precondition_required
        - payload_too_large
        - idempotency_conflict
        - idempotency_in_progress
        - invalid_transition
        - campaign_requirements_missing
        - campaign_read_only
        - primary_owner_required
        - participant_role_invalid
        - participant_exists
        - participant_not_found
        - reference_not_verified
        - artifact_not_owned
        - artifact_not_found
        - artifact_compensation_failed
        - material_type_not_allowed
        - material_too_large
        - material_empty
        - material_exists
        - material_not_found
        - version_conflict
        - invalid_item_transition
        - item_requirements_missing
        - item_content_required
        - item_blocked
        - item_terminal
        - assignee_not_authorized
        - dependency_exists
        - dependency_not_found
        - dependency_self
        - dependency_cycle
        - dependency_campaign_mismatch
        - dependency_terminal
        - artifact_exists
        - artifact_asset_mismatch
        - artifact_link_not_found
        - content_hash_mismatch
        - dependency_unavailable
        - dependency_invalid_response
        - rate_limited
        - internal_error
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, correlationId]
          properties:
            code: { $ref: '#/components/schemas/PublicErrorCode' }
            message: { type: string }
            correlationId: { type: string, minLength: 1, maxLength: 128 }
            details: {}
...
    MaterialAccessEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [url, expiresAt]
          properties:
            url: { type: string, format: uri }
            expiresAt: { type: string, format: date-time }
...
    ContentAssetCreate:
      type: object
      additionalProperties: false
      required: [assetKind, title]
      properties:
        assetKind:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[a-z][a-z0-9_-]*$'
        title: { type: string, minLength: 1, maxLength: 200 }
    ContentAsset:
      type: object
      additionalProperties: false
      required:
        - id
        - itemId
        - campaignId
        - assetKind
        - title
        - currentVersionNumber
        - version
        - createdBy
        - updatedBy
        - createdAt
        - updatedAt
      properties:
        id: { type: string, format: uuid }
        itemId: { type: string, format: uuid }
        campaignId: { type: string, format: uuid }
        assetKind: { type: string, minLength: 1, maxLength: 64 }
        title: { type: string, minLength: 1, maxLength: 200 }
        currentVersionNumber: { type: integer, minimum: 0 }
        version: { type: integer, minimum: 1 }
        createdBy: { type: string, format: uuid }
        updatedBy: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ContentAssetListEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/ContentAsset' }
    ContentAssetMutationEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          allOf:
            - { $ref: '#/components/schemas/ContentAsset' }
            - type: object
              required: [itemVersion]
              properties:
                itemVersion: { type: integer, minimum: 1 }
    ContentVersionCreate:
      type: object
      additionalProperties: false
      required: [body, metadata, freeze]
      properties:
        body: { type: [string, 'null'], maxLength: 1048576 }
        metadata: { type: object, maxProperties: 1000 }
        freeze: { type: boolean }
    ContentVersion:
      type: object
      additionalProperties: false
      required:
        - assetId
        - versionNumber
        - body
        - metadata
        - contentHash
        - createdBy
        - createdAt
        - frozenAt
      properties:
        assetId: { type: string, format: uuid }
        versionNumber: { type: integer, minimum: 1 }
        body: { type: [string, 'null'], maxLength: 1048576 }
        metadata: { type: object }
        contentHash: { type: string, pattern: '^[a-f0-9]{64}$' }
        createdBy: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        frozenAt: { type: [string, 'null'], format: date-time }
    ContentVersionListEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/ContentVersion' }
    ContentVersionMutationEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          allOf:
            - { $ref: '#/components/schemas/ContentVersion' }
            - type: object
              required: [assetVersion]
              properties:
                assetVersion: { type: integer, minimum: 1 }
    ItemArtifactLink:
      type: object
      additionalProperties: false
      required: [artifactId]
      properties:
        artifactId: { type: string, format: uuid }
        assetId: { type: string, format: uuid }
    ItemArtifactRemoval:
      type: object
      additionalProperties: false
      required: [artifactLinkId]
      properties:
        artifactLinkId: { type: string, format: uuid }
    ItemArtifact:
      type: object
      additionalProperties: false
      required:
        - id
        - itemId
        - campaignId
        - assetId
        - artifactId
        - artifactOwnerId
        - filename
        - contentType
        - sizeBytes
        - sha256
        - createdBy
        - createdAt
      properties:
        id: { type: string, format: uuid }
        itemId: { type: string, format: uuid }
        campaignId: { type: string, format: uuid }
        assetId: { type: [string, 'null'], format: uuid }
        artifactId: { type: string, format: uuid }
        artifactOwnerId: { type: string, format: uuid }
        filename: { type: string, minLength: 1, maxLength: 255 }
        contentType: { type: string, minLength: 1, maxLength: 255 }
        sizeBytes: { type: integer, minimum: 1, maximum: 26214400 }
        sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
        createdBy: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
    ItemArtifactListEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/ItemArtifact' }
    ItemArtifactMutationEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [artifact, itemVersion]
          properties:
            artifact: { $ref: '#/components/schemas/ItemArtifact' }
            itemVersion: { type: integer, minimum: 1 }
    ItemArtifactRemovalEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [artifactLinkId, itemVersion]
          properties:
            artifactLinkId: { type: string, format: uuid }
            itemVersion: { type: integer, minimum: 1 }
...